requirements.txt
accelerate==1.9.0
aiobotocore==2.12.1
aiodns==3.1.1
aiofiles==23.2.1
aiohttp==3.9.4
aioitertools==0.11.0
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
asteval==0.9.32
async-timeout==4.0.3
attrs==23.2.0
azure-core==1.28.0
azure-identity==1.13.0
azure-storage-blob==12.17.0
bardapi==1.0.0
beautifulsoup4==4.13.5
bitsandbytes==0.46.1
boto3==1.26.63
botocore==1.29.165
browser-cookie3==0.20.1
cbor==1.0.0
cchardet==2.1.7
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.1.1
colorama==0.4.6
construct==2.10.68
contourpy==1.3.2
coverage==7.4.4
cryptography==44.0.0
cycler==0.11.0
datasets==4.0.0
dill==0.3.8
distro==1.9.0
exceptiongroup==1.2.2
Farama-Notifications==0.0.4
fastapi==0.110.0
filelock==3.15.4
FlagEmbedding==1.3.5
fonttools==4.58.5
frozenlist==1.4.1
fsspec==2024.10.0
GPUtil==1.4.0
gunicorn==22.0.0
gymnasium==0.29.1
h11==0.16.0
h2==4.2.0
hf-xet==1.1.5
hpack==4.1.0
httpcore==1.0.9
httptools==0.6.1
httpx==0.28.1
huggingface-hub==0.33.2
hyperframe==6.1.0
idna==3.6
idna-ssl==1.1.0
ijson==3.4.0
imageio==2.22.2
inscriptis==2.6.0
ir_datasets==0.5.11
isaacsim==4.5.0.0
isaacsim-app==4.5.0.0
isaacsim-asset==4.5.0.0
isaacsim-benchmark==4.5.0.0
isaacsim-code-editor==4.5.0.0
isaacsim-core==4.5.0.0
isaacsim-cortex==4.5.0.0
isaacsim-example==4.5.0.0
isaacsim-gui==4.5.0.0
isaacsim-kernel==4.5.0.0
isaacsim-replicator==4.5.0.0
isaacsim-rl==4.5.0.0
isaacsim-robot==4.5.0.0
isaacsim-robot-motion==4.5.0.0
isaacsim-robot-setup==4.5.0.0
isaacsim-ros1==4.5.0.0
isaacsim-ros2==4.5.0.0
isaacsim-sensor==4.5.0.0
isaacsim-storage==4.5.0.0
isaacsim-template==4.5.0.0
isaacsim-test==4.5.0.0
isaacsim-utils==4.5.0.0
isodate==0.6.1
jeepney==0.9.0
Jinja2==3.1.4
jiter==0.10.0
jmespath==1.0.1
joblib==1.5.1
kiwisolver==1.4.4
llvmlite==0.42.0
loguru==0.7.3
lxml==6.0.1
lz4==4.4.4
MarkupSafe==3.0.2
matplotlib==3.10.3
mpmath==1.3.0
msal==1.23.0
msal-extensions==1.0.0
multidict==6.0.5
multiprocess==0.70.16
nest-asyncio==1.5.6
networkx==3.4.2
numba==0.59.1
numpy==1.26.0
numpy-quaternion==2023.0.3
nvidia-cublas-cu11==11.11.3.6
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu11==9.1.0.70
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu11==10.3.0.86
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu11==11.7.5.86
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-lula-no-cuda==0.10.1
nvidia-nccl-cu11==2.21.5
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu11==11.8.86
nvidia-nvtx-cu12==12.6.77
nvidia-srl-math==0.14.0
nvidia-srl-usd==0.19.1
nvidia-srl-usd-to-urdf==0.7.2
nvidia_srl_base==0.10.0
nvsmi==0.4.2
oauthlib==3.2.2
objaverse==0.1.7
omniverse-kit==106.5.0.162521
openai==1.93.0
opencv-python==4.11.0.86
opencv-python-headless==4.9.0.80
osqp==0.6.5
packaging==23.0
pandas==2.3.1
peft==0.17.1
pickle5==0.0.11
pillow==11.0.0
Pint==0.20.1
plotly==5.3.1
portalocker==2.7.0
protobuf==6.32.0
psutil==5.9.8
pyarrow==21.0.0
pybullet==3.2.7
pycares==4.3.0
pycparser==2.22
pycryptodomex==3.23.0
pydantic==2.9.2
pydantic_core==2.23.4
pyparsing==3.0.9
pyperclip==1.8.0
pypng==0.20220715.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-gemini-api==2.4.12
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
qdldl==0.1.7.post1
qrcode==7.4.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==1.3.1
s3transfer==0.6.1
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.14.0
selenium==4.14.0
sentence-transformers==5.0.0
sentencepiece==0.2.1
sentry-sdk==1.43.0
setuptools-scm==8.0.4
six==1.16.0
sniffio==1.3.1
socksio==1.0.0
soupsieve==2.8
starlette==0.36.3
sympy==1.13.1
threadpoolctl==3.6.0
tokenizers==0.21.4
toml==0.10.2
tomli==2.0.1
torch==2.5.1+cu118
torchaudio==2.5.0+cu118
torchvision==0.20.1+cu118
tornado==6.4.2
tqdm==4.67.1
transformers==4.51.3
trec-car-tools==2.6
trimesh==4.6.13
triton==3.1.0
typing-inspection==0.4.1
typing_extensions==4.11.0
tzdata==2025.2
unlzw3==0.2.3
urllib3==1.26.20
usd-core==23.11
uvicorn==0.29.0
warc3-wet==0.2.5
warc3-wet-clueweb09==0.2.5
watchdog==4.0.0
webbot==0.34
websockets==12.0
wrapt==1.16.0
xxhash==3.5.0
yarl==1.9.4
zlib-state==0.1.9
